+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Please add your entries according to this format.
 ## Unreleased
 - Add support for android projects with the `com.android.kotlin.multiplatform.library` plugin (#422)
 - KtFmt to 0.58
+- Deprecated `manageTrailingCommas` in favor of new `trailingCommaManagementStrategy` property
 
 ##  Version 0.23.0
 - Add `ktfmt.useClassloaderIsolation` property to toggle between processIsolation and classloaderIsolation for the Gradle Worker (#427)